refactor(EnergyBalanceDetailsForPlayerGraph): Rename folder
