+++ packages/client/src/modules/charts/EnergyBalanceCharts/EnergyBalanceDetailsForPlayerGraph.tsx
