+++ packages/client/src/modules/charts/EnergyBalanceCharts/EnergyBalanceForPlayerChart.tsx
